refactor: OrderPrice 패키지 변경
